refactor: separate modules between `core`, `ui` and `shared`
