+++ lib/core/errors/UploadCancelledError.ts
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-import { t } from '../utils/l10n.ts'
+import { t } from '../../shared/utils/l10n.ts'
 
 export class UploadCancelledError extends Error {
